# [Performance] Cache classname collection on DynamicSourceLocatorProvider