--- a/.github/workflows/tests.yaml
+++ b/.github/workflows/tests.yaml
@@ -17,7 +17,7 @@ jobs:
         strategy:
             fail-fast: false
             matrix:
-                os: [ubuntu-latest, macos-12, windows-latest]
+                os: [ubuntu-latest, macos-13, windows-latest]
 
         runs-on: ${{ matrix.os }}
         timeout-minutes: 3