--- a/src/Caching/Enum/CacheKey.php
+++ b/src/Caching/Enum/CacheKey.php
@@ -18,4 +18,9 @@ final class CacheKey
      * @var string
      */
     public const FILE_HASH_KEY = 'file_hash';
+
+    /**
+     * @var string
+     */
+    public const PATHS_HASH_KEY = 'paths_hash';
 }