--- a/src/NodeTypeResolver/Reflection/BetterReflection/SourceLocatorProvider/DynamicSourceLocatorProvider.php
+++ b/src/NodeTypeResolver/Reflection/BetterReflection/SourceLocatorProvider/DynamicSourceLocatorProvider.php
@@ -27,11 +27,6 @@ use Rector\Util\FileHasher;
 final class DynamicSourceLocatorProvider implements ResetableInterface
 {
     /**
-     * @var string
-     */
-    private const CLASSNAMES_CACHE = 'classname_cache';
-
-    /**
      * @var string[]
      */
     private array $filePaths = [];