Continue of PR:

- https://github.com/rectorphp/rector-src/pull/5735

this ensure classname collection on `DynamicSourceLocatorProvider` to be cached, ensure no more `$sourceLocator->locateIdentifiersByType()` if already cached, just call the `$this->reflectionProvider->getClass($classNameCache)`

On parallel process, the `DynamicSourceLocatorDecorator` always new instance due to Worker usage.

**Before** => 56 seconds

```
➜  rector-src git:(main) time bin/rector process --clear-cache 
 2176/2176 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  381.72s user 15.69s system 703% cpu 56.459 total
```

**After**  => 32 seconds.

```
➜  rector-src git:(cache-classname) time bin/rector process --clear-cache
 2176/2176 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  221.64s user 9.05s system 716% cpu 32.190 total
```