@samsonasik 2024-04-26
@TomasVotruba Latest class-leak version seems cause error:

https://github.com/rectorphp/rector-src/actions/runs/8854029950/job/24316165596?pr=5845#step:5:1